Expose fullscreen control/state
Adding a two way `fullscreen` parameter, setting it to `True` puts map in to
full screen mode, and setting it to `False` exits full screen mode. Current
state is observable via this parameter as well.

Note: typically browser will not allow to enter full screen if the event has not
been triggered by the user, so this can only be used in response to the user
action like a button click.

Safari: entering fullscreen mode via custom ui click doesn't work, looks like
safari expects fullscreen request and UI interaction that triggers it to happen
in the same animation frame and not just in "close enough in time". Observing
fullscreen state and exiting it does work however. And it's an important use
case, for example if user completes some input in full screen mode it might be
desirable to exit fullscreen at that point.